pyproject.toml
[tool.poetry]
name = "storytopia-backend"
version = "0.1.0"
description = ""
authors = ["Jason Xiang <xiangchuhang@gmail.com>"]
packages = [{include = "storytopia_backend"}]

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.4.0"
cachecontrol = "0.14.0"
cachetools = "5.3.3"
certifi = "2024.7.4"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
cryptography = "42.0.8"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi = "0.111.0"
fastapi-cli = "0.0.4"
firebase-admin = "6.5.0"
google-api-core = "2.19.1"
google-api-python-client = "2.137.0"
google-auth = "2.32.0"
google-auth-httplib2 = "0.2.0"
google-cloud-core = "2.4.1"
google-cloud-firestore = "2.16.1"
google-cloud-storage = "2.17.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.7.1"
googleapis-common-protos = "1.63.2"
grpcio = "1.65.0"
grpcio-status = "1.62.2"
h11 = "0.14.0"
httpcore = "1.0.5"
httplib2 = "0.22.0"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
jinja2 = "3.1.4"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
msgpack = "1.0.8"
orjson = "3.10.6"
proto-plus = "1.24.0"
protobuf = "4.25.3"
pyasn1 = "0.6.0"
pyasn1-modules = "0.4.0"
pycparser = "2.22"
pydantic = "2.8.2"
pydantic-core = "2.20.1"
pygments = "2.18.0"
pyjwt = "2.8.0"
pyparsing = "3.1.2"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
requests = "2.32.3"
rich = "13.7.1"
rsa = "4.9"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.37.2"
typer = "0.12.3"
typing-extensions = "4.12.2"
ujson = "5.10.0"
uritemplate = "4.1.1"
urllib3 = "2.2.2"
uvicorn = "0.30.1"
uvloop = "0.19.0"
watchfiles = "0.22.0"
websockets = "12.0"
google-cloud-aiplatform = "^1.60.0"
vertexai = "^1.60.0"
asyncio = "^3.4.3"
openai = "^1.40.0"
pytest = "^8.3.2"
pylint = "^3.2.6"
twisted = "^24.3.0"


[tool.poetry.group.dev.dependencies]
anyio = "^4.4.0"
pytest-asyncio = "^0.23.8"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
pytest-twisted = "^1.14.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:Type google._upb._message:DeprecationWarning"
]
